+++ src/carlson/elliprj.rs
@@ -13,7 +13,7 @@
 use std::mem::swap;
 
 use crate::{
-    crate_util::{declare, let_mut},
+    crate_util::{case, check, declare, let_mut, return_if_valid_else},
     elliprc, elliprd, elliprf, StrErr,
 };
 use num_traits::Float;
@@ -42,6 +42,14 @@ use num_traits::Float;
 ///   - any of x, y, or z is negative, or more than one of them are zero,
 ///   - or p = 0.
 /// - Returns the Cauchy principal value if p < 0.
+///
+/// ## Special Cases
+/// - RJ(x, x, x, x) = 1/(x sqrt(x))
+/// - RJ(x, y, z, z) = RD(x, y, z)
+/// - RJ(x, x, x, p) = 3/(x-p) * (RC(x, p) - 1/sqrt(x)) for x ≠ p and xp ≠ 0
+/// - RJ(x, y, y, y) = RD(x, y, y)
+/// - RJ(x, y, y, p) = 3/(p-y) * (RC(x, y) - RC(x, p)) for y ≠ p
+/// - RJ(x, y, z, p) = 0 for x = ∞ or y = ∞ or z = ∞ or p = ∞
 ///
 /// ## Graph
 /// ![Symmetric Elliptic Integral of the Third Kind](https://github.com/p-sira/ellip/blob/main/figures/elliprj_plot.svg?raw=true)
@@ -103,21 +111,17 @@ pub fn elliprj<T: Float>(x: T, y: T, z: T, p: T) -> Result<T, StrErr> {
 #[inline]
 #[numeric_literals::replace_float_literals(T::from(literal).unwrap())]
 fn elliprc1p<T: Float>(y: T) -> Result<T, StrErr> {
-    // We can skip this check since the call from elliprj already did the check.
-    // if y == -1.0 {
-    //     return Err("elliprc1p: y cannot be -1.0.");
-    // }
-
+    // We can skip y = -1 check since the call from elliprj already did the check.
     // for 1 + y < 0, the integral is singular, return Cauchy principal value
-    if y < -1.0 {
-        Ok((1.0 / -y).sqrt() * elliprc(-y, -1.0 - y)?)
+    if y > 0.0 {
+        Ok(y.sqrt().atan() / y.sqrt())
     } else if y == 0.0 {
         Ok(1.0)
-    } else if y > 0.0 {
-        Ok(y.sqrt().atan() / y.sqrt())
     } else if y > -0.5 {
         let arg = (-y).sqrt();
         Ok((arg.ln_1p() - (-arg).ln_1p()) / (2.0 * (-y).sqrt()))
+    } else if y < -1.0 {
+        Ok((1.0 / -y).sqrt() * elliprc(-y, -1.0 - y)?)
     } else {
         Ok(((1.0 + (-y).sqrt()) / (1.0 + y).sqrt()).ln() / (-y).sqrt())
     }
@@ -176,7 +180,7 @@ fn _elliprj<T: Float>(x: T, y: T, z: T, p: T) -> Result<T, StrErr> {
 
     let mut fmn = 1.0;
     let mut rc_sum = 0.0;
-
+    let mut ans = nan!();
     for _ in 0..N_MAX_ITERATION {
         let rx = xn.sqrt();
         let ry = yn.sqrt();
@@ -195,7 +199,6 @@ fn _elliprj<T: Float>(x: T, y: T, z: T, p: T) -> Result<T, StrErr> {
         let lambda = rx * ry + rx * rz + ry * rz;
         an = (an + lambda) / 4.0;
         fmn = fmn / 4.0;
-
         if fmn * q < an {
             // Calculate and return
             let x = fmn * (a0 - x) / an;
@@ -223,7 +226,8 @@ fn _elliprj<T: Float>(x: T, y: T, z: T, p: T) -> Result<T, StrErr> {
                     + 45.0 * e2 * e2 * e3 / 272.0
                     - 9.0 * (e3 * e4 + e2 * e5) / 68.0);
 
-            return Ok(result + 6.0 * rc_sum);
+            ans = result + 6.0 * rc_sum;
+            break;
         }
 
         xn = (xn + lambda) / 4.0;
@@ -233,7 +237,18 @@ fn _elliprj<T: Float>(x: T, y: T, z: T, p: T) -> Result<T, StrErr> {
         delta = delta / 64.0;
     }
 
-    Err("elliprj: Failed to converge.")
+    return_if_valid_else!(ans, {
+        check!(@nan, elliprj, [x, y, z, p]);
+        case!(@any [x, y, z, p] == inf!(), T::zero());
+        eprintln!(
+            "{}, {}, {}, {}",
+            x.to_f64().unwrap(),
+            y.to_f64().unwrap(),
+            z.to_f64().unwrap(),
+            p.to_f64().unwrap()
+        );
+        Err("elliprj: Failed to converge.")
+    })
 }
 
 #[cfg(not(feature = "reduce-iteration"))]
@@ -294,27 +309,40 @@ mod tests {
     }
 
     #[test]
-    fn test_elliprj_err() {
-        // negative argument
+    fn test_elliprj_special_cases() {
+        use std::f64::{INFINITY, NAN};
+        // Negative arguments: should return Err
         assert!(elliprj(-1.0, 1.0, 1.0, 1.0).is_err());
         assert!(elliprj(1.0, -1.0, 1.0, 1.0).is_err());
         assert!(elliprj(1.0, 1.0, -1.0, 1.0).is_err());
-        // more than one zero among x, y, z
+        // More than one zero among x, y, z: should return Err
         assert!(elliprj(0.0, 0.0, 1.0, 1.0).is_err());
         assert!(elliprj(0.0, 1.0, 0.0, 1.0).is_err());
         assert!(elliprj(1.0, 0.0, 0.0, 1.0).is_err());
-        // p == 0
+        // p = 0: should return Err
         assert!(elliprj(1.0, 1.0, 1.0, 0.0).is_err());
+        // NANs: should return Err
+        assert!(elliprj(NAN, 1.0, 1.0, 1.0).is_err());
+        assert!(elliprj(1.0, NAN, 1.0, 1.0).is_err());
+        assert!(elliprj(1.0, 1.0, NAN, 1.0).is_err());
+        assert!(elliprj(1.0, 1.0, 1.0, NAN).is_err());
+        // Infs: should return zero
+        assert_eq!(elliprj(INFINITY, 1.0, 1.0, 1.0).unwrap(), 0.0);
+        assert_eq!(elliprj(1.0, INFINITY, 1.0, 1.0).unwrap(), 0.0);
+        assert_eq!(elliprj(1.0, 1.0, INFINITY, 1.0).unwrap(), 0.0);
+        assert_eq!(elliprj(1.0, 1.0, 1.0, INFINITY).unwrap(), 0.0);
+    }
+
+    #[test]
+    fn cover_deadcode() {
+        // else branch
+        assert!(elliprc1p(-0.6).unwrap().is_finite());
+        // y < -1
+        assert!(elliprc1p(-1.1).unwrap().is_finite());
     }
 }
 
 #[cfg(feature = "reduce-iteration")]
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn force_fail_to_converge() {
-        assert!(elliprj(0.2, 0.5, 1e300, 1.0).is_err());
-    }
+crate::test_force_unreachable! {
+    assert!(elliprj(0.2, 0.5, 1e300, 1.0).is_err());
 }